moving ok button validator into validators file
